backend/pyproject.toml
[tool.poetry]
name = "shops-mediasoft"
version = "0.1.0"
description = ""
authors = ["NikSen <niksen09@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
